# Conflicts:
#	source/UI/Victory_Select.cpp
#	source/level/TutorialLevel.cpp

+++ source/system/CameraSystem.cpp
@@ -28,11 +28,11 @@ namespace Camera
 
 			AEInputGetCursorPosition(&mousex, &mousey);
 
-			mousex = mousex - f32(AEGfxGetWindowWidth()) * 0.5f;
-			mousey = -mousey + f32(AEGfxGetWindowHeight()) * 0.5f;
+			mousex = mousex - s32(f32(AEGfxGetWindowWidth()) * 0.5f);
+			mousey = -mousey + s32(f32(AEGfxGetWindowHeight()) * 0.5f);
 
-			mousex = AEClamp(mousex, AEGfxGetWinMinX(), AEGfxGetWinMaxX());
-			mousey = AEClamp(mousey, AEGfxGetWinMinY(), AEGfxGetWinMaxY());
+			mousex = AEClamp(mousex, (s32)AEGfxGetWinMinX(), (s32)AEGfxGetWinMaxX());
+			mousey = AEClamp(mousey, (s32)AEGfxGetWinMinY(), (s32)AEGfxGetWinMaxY());
 
 			mouse.x = mousex;
 			mouse.y = mousey;
@@ -61,11 +61,11 @@ namespace Camera
 
 			AEInputGetCursorPosition(&mousex, &mousey);
 
-			mousex = mousex - f32(AEGfxGetWindowWidth()) * 0.5f;
-			mousey = -mousey + f32(AEGfxGetWindowHeight()) * 0.5f;
+			mousex =  mousex - s32(f32(AEGfxGetWindowWidth()) * 0.5f);
+			mousey = -mousey + s32(f32(AEGfxGetWindowHeight()) * 0.5f);
 
-			mousex = AEClamp(mousex, AEGfxGetWinMinX(), AEGfxGetWinMaxX());
-			mousey = AEClamp(mousey, AEGfxGetWinMinY(), AEGfxGetWinMaxY());
+			mousex = AEClamp((f32)mousex, AEGfxGetWinMinX(), AEGfxGetWinMaxX());
+			mousey = AEClamp((f32)mousey, AEGfxGetWinMinY(), AEGfxGetWinMaxY());
 
 			cam.buffer_x = f32(transform->pos.x) + mousex;
 			cam.buffer_y = f32(transform->pos.y) + mousey;
